[#main] 몬스터 변경 및 UI적용

+++ Assets/3-Scripts/DataScripts/Monster/MonsterData.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "MonsterData/MonsterData")]
 public class MonsterData : ScriptableObject
 {
-    public int monsterSpeed;
+    public float monsterSpeed;
     public int maxHP;
     public int contectDamage;
     public int attackDamage;
